client/requirements.txt
pyserial
requests
python-dotenv
